Embed upspin config in code!
Instead of relying on a configuration on the local filesystem, the
upspin configuration is now embedded in the code.

I create a read-only upspin account (no DirEndpoint nor StoreEndpoint),
anonymous@yopmail.com, which can be used by anybody for reading public
files on upspin. The embedded configuration uses that account